Core image processing for coil-element SNR.
Previous circle-finding technique failed due to non-uniform phantoms.

Here I have implemented the "triangle method" for thresholding the
images, so as to find the centre of the phantom (using the centre of
mass).

A 20x20 pixel ROI is then placed midway between the brightest part of
the image and the phantom centre and the smoothed-subtraction method is
used to determine the SNR in this region.

Initial testing shows good agreement with manual tests performed with
ImageJ.