Routine MR QA analysis script.
This script reads the csv file generated by the mr_coil_qa script.

It then sorts the data, collating it into groups with unique
combinations of coil name, element id and slice location. The measured
SNR for each group is then plotted against the date of the tests and
these plots are all saved in a storage folder.

Saving the plots is really to help with project development. This script
could form the basis of a more involved data analysis algorithm, which
could try to choose the best data (highest SNR location for each
element) and track performance over time (setting baselines and
e-mailing out-of-tolerance alerts).