Implement ShapeDetector for all tasks
There was some duplication of the same object finding tasks for
the phantom. Sometimes it was finding a circle for axial view or
other shapes. This ShapeDetector class in the tools.py package
should be the go-to for any detection of objects.

Due to the slightly different method of finding circles, it meant
that the tests had to be updated to take into account slightly
different values for signal. The testing needs to be made more
robust to these kinds of changes eventually.