Trigger processing tasks and show results
There were a few minor bugs that were preventing this, such as
not converting UUIDs to string etc. Also, have begun work on
docker-compose to make sure that all components are alive
when testing or deploying, such as RabbitMQ and Celery.

Resetting password doesn't work because there isn't proper mail
server integration. This needs to be fixed because I've had to
create multiple new users as I've forgotten passwords.

Deleted the followers table as it really is not relevant right now
and is a hangover from when I followed the initial Flask tutorial
to build the initial web app.

Updated opencv version as Docker couldn't find the version when
trying to build through docker-compose.