Add Toshiba support for SNR. Force reading of DICOM
Some support for Toshiba files has been added but this has not
been properly tested. Currently only SNR test is known to work
and even then no test data has been added to repo.

Also changed interface design so that main functions receive list
of pydicom objects and not list of filepaths, that way we can be more
robust in reading in files.