Improve phantom circle and MTF square finding
The max_radius possible for circle finding has been reduced to
200 pixels, with the phantom size known to be less than that on
images of size 512x512.

The find_square method for finding the MTF square didn't order
the points of the squares it returned, meaning the calculations
were not reliable. It also now takes into account when the square
is tilted forwards and backwards.

Also improved results when there are multiple input dicoms for
whatever reason. Wrapped around try except in case one of the
dicoms in the directory are bad, which was the case for a test
case from the Broomfield Philips data set.

Resolves: #18