Improve circle and square finding for different matrices
Changed minimum radius in HoughCircles and also the MTF
block size to take into account small radius from 256 matrix
size (smaller matrix sizes will be an issue).

Fixed MTF block corner ordering to go from top-right anti-
clockwise.

Also the get_roi function swaps x, y to go column-major - it
is not understood why this has to be column major.

Fixed get_top_edge_vector_and_centre and the equivalent for
the right edge, but the vector returned is nonsense. This is ok
for now because the vector is no longer used, but would be good
to factor this out or fix this in the future.

Improved tests and added tests cases from East Kent and Broomfield
scanners as the MTF tilt and different matrix sizes were presenting
issues. Would be good to consolidate tests at some point.

Resolves: #7, #16