Accomodate ghosting phantom being off edge of image
When the phantom is off the edge of the image the eligible area
was inappropriately allowing potential ghost ROIs to be off the edge
of the image, leading to an exception being raised. I also removed
the 1 pixel padding around the signal bounding box and this provided
no extra value but created issue about tracking where edge of the
phantom was.

Also reduced the threshold value for finding the phantom due to large
phantom non-uniformity meaning bounding box was cutting off phantom.

Issue 23 mistakenly mentions this may be due to a square phantom but
this method is agnostic to phantom shape as long as there is one
contiguous phantom.

Resolves: #23