HZN-38 Make report path keyword arg not positional
The SNR reporting was broken when not specifying the slice width
because the report variable was positional and being asigned to the
'measured_slice_width' variable and the report_path variable was using
the default value of false. By using a keyword argument for the
report_path this should not longer be an issue.