Modified dockerfile so that Hazen can be run in container
testing two different ways (currently method 1 is commented out)

method 1 - run as interactive container, this way is similar to how
 hazen is currently run in a venv, where the user must activate the
container and can then run the hazen command line tool as normal.
the downside might be that it keeps some of the confusion about using
venvs i.e. that there is a step to perform before using hazen.

method 2 - run as containerised application. running the app this way
is the simplest way to do it and removes almost all the user interaction
with docker. can create an alias function where 'hazen' can run the
docker run .... etc function that mounts current working directory etc.
this option seems like the best choice to me at the moment, but the
it does mean that autocomplete functions won't work - not sure if thats
a big deal but is a bit of hassle.