Added hazen-app shell script and tidied dockerfile
Added hazen-app bash shell script as a wrapper for running the docker command that starts the hazen app container and mounts volumes etc. This should make it much easier for someone with little knowledge of implementation to run the application. 

Idea is that the Hazen work instruction will have a list of copy/paste commands that will retrieve this script from the repo and move it to /usr/bin. They should then be able to run 'hazen-app' in the terminal and the docker application will run with the current directory mounted.  

If implementation is sound, we can add checks in this shell script, such as to ensure docker image is up to date or display warnings if not.