added bitbucket build instructions for docker and improved app script
Updated bitbucket pipelines so it builds docker image, sends new version with :latest tag and add :build_number tag and pushes that too.

Updated hazen-app so that it checks for new version of hazen on :latest tag and pulls new version if available. Unable to control version update without enterprise version of docker, so taking approach of forcing user to update when application is run. 

Added command line argument to hazen-app so that user can request specific tagged version of hazen image on dockerhub