HZN-77 Change default output to single number and add 'verbose' flag
A new single value is returned by default. The function calculates
the T1 or T2 relaxometry time for each of the 14 test spheres, and
for the background water. The fractional difference between the
measured and published relaxometry time is calculated for each
sphere. The RMS of these fractional differences is returned.

Using the 'verbose' flag will return the extended output information.