creating test functions for hazenlib.__init__
I added the first test function for hazenlib __init__. It checks if a file is an enhanced dicom. In the test data I could only find files with the class: 1.2.840.10008.5.1.4.1.1.4 and no file with the class 1.2.840.10008.5.1.4.1.1.4.1 so I only tested part of the conditional statement. I will gradually add more tests functions. This pull request is not complete.