wrote new function to find LSF
As mentioned, I could not find the literature to support the Maivis equation which finds the LSF. However it seems the LSF can be simply calculated by taking the derivative of the ESF http://sciencewise.info/resource/line_spread_function/Line_spread_function_by_Wikipedia. So I changed the function directly. Let me know if this fix is ok.