updated hazenlib_init_ tests
Hi @laurencejackson, I have added the changes you suggested (If my interpretation was correct). I could not add the test_field_of_view with the other tests because the GE manufacturer was not working with the hazenlib function. 
Also, I could not add the enhanced_dicom = True bit because I do not have a test image which returns true for that condition.