fixes poor thresholding on shape detector
The shape detector was performing poorly due to suboptimal thresholding values. The previous value of MAX_PIXEL // 5 was too inflexible.

This PR updates the tools.ShapeDetector:find_contours method to use a the Li method for automatic thresholding.

This resulted in different values for the slice position test so the test values needed to be changed. On inspection the new values looked like a better fit to the data.