Updated test_cli
Also removed hazenlib/relaxometry.py (replaced by hazenlib/tests/relaxometry.py)

The CLI for relaxometry does not work on a standard installation via pip. It requires the hazenlib/tasks/data/relaxometry/ folder, which is not installed. I've tried a few things (including MANIFEST.in), but don't know how to get this folder, along with the non-python files it contains, to install to the hazenlib library.