Refactoring of ACR_Uniformity
Added ACR_Tools functionality to uniformity calculation. <1% changes in test uniformity results owing to new thresholding technique, which is more robust to intensity gradients across imaging. The difference in output PIU is due to the fact that the new edge filtering method does not provide an identical centroid as the former centre-of-mass method.