Change mask threshold
On slice 1 the mask was being generated incorrectly on the GE data (it was extending outside the phantom). Some pixels from outside the phantom had signal value above the masking threshold, so these were being picked up and incorporated in the mask. I increased the masking threshold from 0.05 to 0.07 and this meant that the background pixels were no longer being picked up.