test: Ensured backwards compatability with toml loading
In order to preserve backwards comparability with Python 3.9 and 3.10
we try and load tomllib (preferred) and fall back to loading toml
which is currently a development dependency.
Once Python 3.10 support is dropped (October 2026) then toml can be
dropped as a development dependency and this code can be streamlined.

Related to #16